Move shared code into a packlet
