+++ apps/lockfile-explorer-web/package.json
@@ -11,6 +11,16 @@
     "_phase:build": "heft run --only build -- --clean",
     "_phase:test": "heft run --only test -- --clean"
   },
+  "exports": {
+    "./packlets/lfx-shared": "./lib-commonjs/packlets/lfx-shared/index.js"
+  },
+  "typesVersions": {
+    "*": {
+      "packlets/lfx-shared": [
+        "lib/packlets/lfx-shared/index.d.ts"
+      ]
+    }
+  },
   "dependencies": {
     "@lifaon/path": "~2.1.0",
     "@reduxjs/toolkit": "~1.8.6",
