+++ apps/lockfile-explorer-web/src/helpers/localStorage.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { type LockfileEntry, LockfileEntryFilter } from '../parsing/LfxGraph';
+import { type LockfileEntry, LockfileEntryFilter } from '../packlets/lfx-shared';
 
 const BOOKMARK_KEY: string = 'LOCKFILE_EXPLORER_BOOKMARKS';
