+++ apps/lockfile-explorer-web/src/packlets/lfx-shared/IAppContext.ts
