+++ apps/lockfile-explorer-web/src/packlets/lfx-shared/IJsonLfxGraph.ts
