+++ apps/lockfile-explorer-web/src/packlets/lfx-shared/index.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+// See LICENSE in the project root for license information.
+
+export * from './IAppContext';
+export * from './IJsonLfxGraph';
+export * from './IJsonLfxWorkspace';
+export * from './LfxGraph';
+export * as lfxGraphSerializer from './lfxGraphSerializer';
