+++ apps/lockfile-explorer-web/src/parsing/readLockfile.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import type { LockfileEntry } from './LfxGraph';
+import type { LockfileEntry } from '../packlets/lfx-shared';
 import * as lfxGraphLoader from './lfxGraphLoader';
 
 const serviceUrl: string = window.appContext.serviceUrl;
