+++ apps/lockfile-explorer-web/src/types/AppContext.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import type { IAppContext } from './IAppContext';
+import type { IAppContext } from '../packlets/lfx-shared';
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/naming-convention
