+++ apps/lockfile-explorer/src/state/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import type { ILfxWorkspace } from '@rushstack/lockfile-explorer-web/lib/types/lfxProtocol';
+import { type IJsonLfxWorkspace } from '@rushstack/lockfile-explorer-web/packlets/lfx-shared';
 
 export interface IAppState {
   lockfileExplorerProjectRoot: string;
@@ -11,5 +11,5 @@ export interface IAppState {
   pnpmfileLocation: string;
   appVersion: string;
   debugMode: boolean;
-  lfxWorkspace: ILfxWorkspace;
+  lfxWorkspace: IJsonLfxWorkspace;
 }
